requirements.txt
aiohttp==3.5.4
mh_z19==0.3.9
aiohttp_jinja2==1.1.1
Jinja2==2.10.1
aiosqlite==0.10.0
Adafruit_DHT==1.4.0
typing==3.6.6
PyYAML==5.1.1